nl_NL.po:add Dutch translation test
To test, I translated one line in Dutch, and it works. :)

```
andy@oceanus:~/src/bric/install_test/bin$ LANG=nl_NL.utf8 ./bric
Gebruik: bric <bestandsnaam>
```

This not only requires the translation in bric, but for the user to have
the locale installed on his system, if I'm not mistaken.

I already had the This locale installed. to install on Debian Linux

https://wiki.debian.org/Locale#Manually