tagstack.c:fix curr warning
This fixes the warning

```
tagstack.c: In function ‘pop’:
tagstack.c:37:10: warning: ‘curr’ may be used uninitialized in this
function [-Wmaybe-uninitialized]
curr->p = NULL;
```

https://travis-ci.com/shnupta/bric/jobs/143067788#L512

I'm not sure if this is a clean fix. I'm not yet familiar enough with
bric to test it properly (and the use of calloc() is still a bit
mysterious to me). The only things I did to test it was to build it
(warning went away), then open a file with bric.