remove malloc, init node
This gets rid of the warning. I change 'curr' to 'node' since we're
dealing with a "Linked List" here.

I don't see that calloc or malloc is needed. I'm initializing `node'
with `foo` after foo is initialized. Before, the warning was happening
because of the node->p = NULL assignment; if the while loop never
executes, node is not initialized before that assignment.