tagstack.c:fix curr warning (#132)
* tagstack.c:fix curr warning

This fixes the warning

```
tagstack.c: In function ‘pop’:
tagstack.c:37:10: warning: ‘curr’ may be used uninitialized in this
function [-Wmaybe-uninitialized]
curr->p = NULL;
```

https://travis-ci.com/shnupta/bric/jobs/143067788#L512

This gets rid of the warning. I changed 'curr' to 'node' since we're
dealing with a "Linked List" here.

I don't see that calloc or malloc is needed. I'm initializing `node'
with `foo` after foo is initialized. Before, the warning was happening
because of the node->p = NULL assignment; if the while loop never
executes, node is not initialized before that assignment.

* tags document:updated with accurate info

(related to #139)